+++ client/app/src/main/java/com/example/witt/data/repository/UserRepositoryImpl.kt
@@ -1,20 +1,20 @@
 package com.example.witt.data.repository
 
-import com.example.witt.data.model.user.request.DuplicateEmailRequest
+import com.example.witt.data.model.remote.user.request.DuplicateEmailRequest
 import com.example.witt.data.mapper.toDuplicateEmailModel
 import com.example.witt.data.mapper.toProfileUploadModel
 import com.example.witt.data.mapper.toUserProfileModel
 import com.example.witt.data.model.local.UserProfile
-import com.example.witt.data.model.profile.request.ProfileUploadRequest
-import com.example.witt.data.model.user.response.toGetUserInfoModel
+import com.example.witt.data.model.remote.user.request.ProfileUploadRequest
+import com.example.witt.data.model.remote.user.response.toGetUserInfoModel
 import com.example.witt.data.source.local.user_profile.ProfileDataSource
-import com.example.witt.data.source.remote.duplicate_check.DuplicateEmailDataSource
-import com.example.witt.data.source.remote.profile.ProfileUploadDataSource
+import com.example.witt.data.source.remote.signup.duplicate_check.DuplicateEmailDataSource
+import com.example.witt.data.source.remote.user.profile.ProfileUploadDataSource
 import com.example.witt.data.source.remote.user.GetUserInfoDataSource
-import com.example.witt.domain.model.profile.remote.ProfileUploadModel
-import com.example.witt.domain.model.user.DuplicateEmailModel
-import com.example.witt.domain.model.user.GetUserInfoModel
-import com.example.witt.domain.model.user.UserProfileModel
+import com.example.witt.domain.model.remote.user.ProfileUploadModel
+import com.example.witt.domain.model.remote.signup.DuplicateEmailModel
+import com.example.witt.domain.model.remote.user.GetUserInfoModel
+import com.example.witt.domain.model.remote.user.UserProfileModel
 import com.example.witt.domain.repository.UserRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
