+++ client/app/src/main/java/com/example/witt/domain/model/remote/user/ProfileUploadModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.profile.remote
+package com.example.witt.domain.model.remote.user
 
 data class ProfileUploadModel (
     val status: Boolean,
