+++ client/app/src/main/java/com/example/witt/domain/use_case/remote/UploadRemoteProfile.kt
@@ -1,6 +1,6 @@
 package com.example.witt.domain.use_case.remote
 
-import com.example.witt.domain.model.profile.remote.ProfileUploadModel
+import com.example.witt.domain.model.remote.user.ProfileUploadModel
 import com.example.witt.domain.repository.UserRepository
 import java.io.File
 import javax.inject.Inject
